Start of a major rework of FAQ 5, turning it from being about building just a custom kernel to being about rebuilding the entire system from source.  ok kjell@
I know I'm going to get a lot of comments about the spelling of
"flavor" vs. "flavour".  Personally, I would not mind having
British/Canadian spellings in the FAQ at all, but here's the problem:
If I use the British/Canadian spelling, in a little while, I'm going to
be sent a diff with all kinds of good wording/spelling fixes by someone,
AND "flavour" changed to "flavor".  Now, I'll either have to commit it,
reject it, or manually extract the offending 10%.  I'll be lazy, and just
commit it.  Then, I'll get notes from a bunch of people pointing out that
"flavour" *is* a valid spelling.

So, sorry.  It goes in as "flavor" with no intended offense.