I publically asked Wim a simple but difficult question about the accounting of donations -- he never answered it.
The next morning he sends me a bill for some hardware he had given me to
work on the Openmoko port, so I sent it back. If someone in the
community has a Freerunner they could spare, I'm still willing to work on
it in what spare time I have.