reorganize 5.3 - building from source:
simply explain the preparation steps and refer to release(8) instead
of duplicating its contents in a much more verbose fashion.

similarly, don't repeat all the instructions from anoncvs.html.
refer to it, keep the tips on running cvs without root and stick to
doing -current as an example.

move the cvs tips to the end of the page for now.