Remove my request for Atheros AR5008E-3NX devices.
One such device was donated some time ago, thanks to Gerrit Meyerheim.
Additional thanks to everyone else who subsequently offered to obtain
such a device for me.

The card I got cannot yet work with our athn(4) driver because it does not
have an EEPROM (which, among other things, provides the PCI device ID).

I now need help with finding a generic EEPROM and/or a way to generate
EEPOM images which contain suitable data. Since this is a narrowly scoped
task which does not require kernel driver development skills, I want to
delegate it. If you are interested in working on this task, and especially
if you have contacts in the OpenWRT community (who have solved this issue
in their OS), let me know.