discontinue sgi.
(On the o350 (where builds happened for years) there is a dma / disk
controller bug which was never fully diagnosed and fixed in the driver.
it locks up the vfs layer after a while, and sometimes causes quite
catastrophic filesystem layout.  I have tired of repairing these situation.
The octeon platform is becoming more viable as a mips64 canarie, so we'll
carry on using it)