Mention crashes on T4-2 machines with older hardware
Koakuma on bugs@ reports that the following firmware versions misbehave:

-> show /HOST hypervisor_version obp_version sysfw_version

  /HOST
    Properties:
         hypervisor_version = Hypervisor 1.15.16 2018/11/28 07:41
         obp_version = OpenBoot 4.38.16 2018/11/28 07:24
         sysfw_version = Sun System Firmware 8.9.11 2018/11/28 07:59

SysFW 8.9.11 (152476-09) however is known to work, containing:

  /HOST
    Properties:
         hypervisor_version = Hypervisor 1.15.16 2018/11/28 07:41
         obp_version = OpenBoot 4.38.16 2018/11/28 07:24
         sysfw_version = Sun System Firmware 8.9.11 2018/11/28 07:59


Since there seems to be no public information available linking SysFW
versions to release dates, this is the we can do so far to narrow down the
range of broken/working firmware versions.

Users with different versions are encouraged to report their state and help
update this information.

OK kettenis