The entire multibooting section needs work and testing.
Reportedly, the GRUB 2 instructions don't work with GPT.
Daniel Blevins invested a lot of time but only got rEFInd to work.
The Windows section is probably outdated as well.

Unless somebody steps up to confirm that things are ok as they
are or to update and fix them, the section will be removed.

discussed with tj