Split the "general" section because it was of excessive size, a plan that benno@ liked in general.
I mostly reordered lines, but dropped two entries that detailed
minor tweaks to a subsystems that was only introduced in this release
in the first place (kstat -w, kstat filtering).  I also deleted a
bogus reference to kstat(8), which does not appear to exist.

Besides, i moved three lines to the ARM section, and i moved
the installboot and FFS2 stuff to the installer section.