various faq/ports/guide.html tweaks;
- introduce a distinction between patches to fix things on OpenBSD
(which generally should go upstream if possible) and those which
are just for ports policy (which shouldn't). for non-maintainers
give a clue that they should talk to the maintainer before
feeding things back themselves in case the maintainer has a good
reason for not pushing diffs (maybe they know they are not quite
right, maybe they're meant to be temporary to work around a
problem in the OS, etc).

- add a few other things to watch out for that come up frequently
(distfile names with just a version number, manual changes to
rcsid lines, missing @sample for new examples files)

- other small tweaks

solene@ landry@ tb@ jca@ like it, typo-spotting by jca