Delete the entry for Kihaguru Njenga Gathura.
They requested an update of their email address but submitted a
malformed entry that looked as if they were struggling to correctly
send plain-text email.  The M line included *two* email addresses
neither of which works; The primary MX for one of the domains rejects
the specified info@ address saying "550 #5.1.0 Address rejected";
the other domain does not even have SOA nor MX records.  The web
server for the (existing) domain redirects HTTP to HTTPs but when
connecting on the HTTPs port, it closes the connection right after
the TLS handshake is completed without waiting for a HTTP request.

A request for repairing the servers and clarifying what exactly
they want listed got no reply within four days.  Also, some time
ago, they asked a very basic question about C programming on misc@
that looked as if they don't understand at all how arrays work in C
and what an array overflow is.

While we do not usually attempt to judge the qualification of people
who want to be listed on support.html, i think it does make sense to
delete entries when very serious doubts arise by chance regarding the
qualification.  I don't think we should advertise a person claiming
"speciality in web applications ... and C programming language" when
they appear to be struggling to send mail, run mail and web servers,
and understand the most basic aspects of C.